add edit to unique vars in key file update
The original key file contained numerous duplicate vars. This is a huge problem as those vars are used for units conversion among other functions, and absolutely must be unique. The updated script changes duplicated values to a new, unique name that Derek Pierson created in a Google Sheet in the SOM Google Drive env.